chore: split out files and use PascalCase
